refactor(site): move workspace schedule controls to its own component
