+++ site/src/pages/WorkspacePage/Workspace.tsx
@@ -37,12 +37,6 @@ export type WorkspaceError =
 export type WorkspaceErrors = Partial<Record<WorkspaceError, unknown>>;
 
 export interface WorkspaceProps {
-  scheduleProps: {
-    onDeadlinePlus: (hours: number) => void;
-    onDeadlineMinus: (hours: number) => void;
-    maxDeadlineIncrease: number;
-    maxDeadlineDecrease: number;
-  };
   handleStart: (buildParameters?: TypesGen.WorkspaceBuildParameter[]) => void;
   handleStop: () => void;
   handleRestart: (buildParameters?: TypesGen.WorkspaceBuildParameter[]) => void;
@@ -81,7 +75,6 @@ export interface WorkspaceProps {
  * Workspace is the top-level component for viewing an individual workspace
  */
 export const Workspace: FC<React.PropsWithChildren<WorkspaceProps>> = ({
-  scheduleProps,
   handleStart,
   handleStop,
   handleRestart,
@@ -186,10 +179,6 @@ export const Workspace: FC<React.PropsWithChildren<WorkspaceProps>> = ({
           workspace={workspace}
           handleUpdate={handleUpdate}
           canUpdateWorkspace={canUpdateWorkspace}
-          maxDeadlineDecrease={scheduleProps.maxDeadlineDecrease}
-          maxDeadlineIncrease={scheduleProps.maxDeadlineIncrease}
-          onDeadlineMinus={scheduleProps.onDeadlineMinus}
-          onDeadlinePlus={scheduleProps.onDeadlinePlus}
         />
 
         {canUpdateWorkspace && (
